Update ethics.md
Made plain language changes as per issue #190 that had previously been reverted in this commit: https://github.com/18F/ux-guide/commit/321e157dcb594a7be3413f9a84162585d0717e0f. Also changed the verb in the phrase "collect informed consent" to "obtain" because it is more accurate to the action (we may or may not obtain it; collecting suggests it's there, we just have to get it).