Updated copy
Copy ready for review. Added some notes and links about 508 and accessibility. I've reached out to folks over slack to see if we've used the IAA with an agency yet. The money is set aside but no research has been conducted and paid for through it. We should come back to the compensation section after we have more examples of how people are paying for recruitment and if we get new guidance from OGC.