Updated consent for Gdoc links (#366)
* Updated consent for Gdoc links

This updates a link that was pointing to an archived English language version w/ a link to the current template; The link to the Spanish language gdoc was actually pointing to a deprecated consent form in google forms in English. I updated this to point to the correct Spanish language example template in Gdocs.

* Replaced link to archive V w/ link to current V

I accidentally repeated the outdated link in my initial version of the PR. Oops! This commit correctly replaces the link to archived version of the English language form with a link to the current version.

---------

Co-authored-by: Ben P <15329140+bpdesigns@users.noreply.github.com>